refactor: Tests have been reorganized
