+++ src/test/data/movieSlice.test.js
@@ -1,5 +1,5 @@
-import moviesSlice, { fetchAllMovies } from '../data/moviesSlice'
-import { moviesMock } from './movies.mocks'
+import moviesSlice, { fetchAllMovies } from '../../data/moviesSlice'
+import { moviesMock } from '../utils/movies.mocks'
 
 describe('MovieSlice test', () => {
   it('should set loading true while action is pending', () => {
