+++ src/test/data/watchLaterSlice.test.js
@@ -1,5 +1,5 @@
-import watchLaterSlice from '../data/watchLaterSlice'
-import { moviesMock } from './movies.mocks'
+import watchLaterSlice from '../../data/watchLaterSlice'
+import { moviesMock } from '../utils/movies.mocks'
 
 describe('watchLaterSlice test', () => {
   const state = { watchLaterMovies: [] }
