+++ src/test/utils/utils.js
@@ -1,34 +1,37 @@
-import React from 'react';
-import { render } from '@testing-library/react';
+import React from 'react'
+import { render } from '@testing-library/react'
 import { Provider } from 'react-redux'
 import { BrowserRouter } from 'react-router-dom'
 import { configureStore } from '@reduxjs/toolkit'
 import { setupListeners } from '@reduxjs/toolkit/dist/query'
-import moviesSlice from '../data/moviesSlice'
-import starredSlice from '../data/starredSlice'
-import watchLaterSlice from '../data/watchLaterSlice'
+import moviesSlice from '../../data/moviesSlice'
+import starredSlice from '../../data/starredSlice'
+import watchLaterSlice from '../../data/watchLaterSlice'
 
 export function renderWithProviders(
   ui,
   {
     preloadedState = {},
     store = configureStore({
-      reducer: { 
-        movies: moviesSlice.reducer, 
+      reducer: {
+        movies: moviesSlice.reducer,
         starred: starredSlice.reducer,
         watchLater: watchLaterSlice.reducer
       },
-      preloadedState,
+      preloadedState
     }),
     ...renderOptions
   } = {}
 ) {
-
   setupListeners(store.dispatch)
 
   function Wrapper({ children }) {
-    return <Provider store={store}><BrowserRouter>{children}</BrowserRouter></Provider>;
+    return (
+      <Provider store={store}>
+        <BrowserRouter>{children}</BrowserRouter>
+      </Provider>
+    )
   }
 
-  return { store, ...render(ui, { wrapper: Wrapper, ...renderOptions }) };
+  return { store, ...render(ui, { wrapper: Wrapper, ...renderOptions }) }
 }
