+++ src/test/views/watchLater.test.js
@@ -1,0 +1,18 @@
+import { screen, waitFor } from '@testing-library/react'
+import userEvent from '@testing-library/user-event'
+import { renderWithProviders } from '../utils/utils'
+import App from '../../App'
+
+it('Watch Later movies page', async () => {
+  renderWithProviders(<App />)
+
+  await userEvent.type(screen.getByTestId('search-movies'), 'forrest gump')
+  await waitFor(() => {
+    expect(screen.getAllByText('Through the Eyes of Forrest Gump')[0]).toBeInTheDocument()
+  })
+  const watchLaterLink = screen.getAllByTestId('watch-later')[0]
+  await waitFor(() => {
+    expect(watchLaterLink).toBeInTheDocument()
+  })
+  await userEvent.click(watchLaterLink)
+})
