--- src/test/watchLater.test.js
@@ -1,24 +1,0 @@
-import { screen, waitFor } from '@testing-library/react'
-import userEvent from '@testing-library/user-event'
-import { renderWithProviders } from './utils'
-import App from '../App'
-
-it('Watch Later movies page', async () => {
-    renderWithProviders(<App />)
-
-    await userEvent.type(screen.getByTestId('search-movies'), 'forrest gump')
-    await waitFor(() => {
-      expect(screen.getAllByText('Through the Eyes of Forrest Gump')[0]).toBeInTheDocument()
-    })
-    const watchLaterLink = screen.getAllByTestId('watch-later')[0]
-    await waitFor(() => {
-        expect(watchLaterLink).toBeInTheDocument()
-    })
-    await userEvent.click(watchLaterLink)
-
-    // const watchLaterink = screen.getByTestId('watch-later-div')
-    // await waitFor(() => {
-    //     expect(watchLaterink).toBeInTheDocument()
-    // })    
-    // await userEvent.click(watchLaterink)
-})
